+++ packages/merkleize-metadata/src/get-accessible-types.ts
@@ -1,8 +1,8 @@
-import { V15 } from "@polkadot-api/substrate-bindings"
+import { UnifiedMetadata } from "@polkadot-api/substrate-bindings"
 import { LookupValue } from "./codecs"
 
 export const getAccessibleTypes = (
-  metadata: V15,
+  metadata: UnifiedMetadata<15 | 16>,
   definitions: Map<number, LookupValue>,
 ): Map<number, number> => {
   const types = new Set<number>()