+++ packages/merkleize-metadata/src/get-metadata.ts
@@ -1,14 +1,17 @@
 import {
   decAnyMetadata,
+  UnifiedMetadata,
   type HexString,
-  type V15,
+  unifyMetadata,
 } from "@polkadot-api/substrate-bindings"
 
-export const getMetadata = (input: Uint8Array | HexString): V15 => {
+export const getMetadata = (
+  input: Uint8Array | HexString,
+): UnifiedMetadata<15 | 16> => {
   try {
-    const { metadata } = decAnyMetadata(input)
-    if (metadata.tag !== "v15") throw new Error("Wrong metadata version")
-    return metadata.value
+    const metadata = unifyMetadata(decAnyMetadata(input))
+    if (metadata.version <= 14) throw new Error("Wrong metadata version")
+    return metadata as UnifiedMetadata<15 | 16>
   } catch (e) {
     throw e || new Error("Unable to decode metadata")
   }