+++ packages/view-builder/tests/view-builder.spec.ts
@@ -2,11 +2,12 @@ import { getKsmMetadata } from "@polkadot-api/metadata-fixtures"
 import { expect, describe, it, beforeAll } from "vitest"
 import { getViewBuilder } from "@/."
 import { getLookupFn } from "@polkadot-api/metadata-builders"
+import { unifyMetadata } from "@polkadot-api/substrate-bindings"
 
 describe("getViewBuilder", () => {
   let builder: ReturnType<typeof getViewBuilder>
   beforeAll(async () => {
-    builder = getViewBuilder(getLookupFn(await getKsmMetadata()))
+    builder = getViewBuilder(getLookupFn(unifyMetadata(await getKsmMetadata())))
   })
 
   it("batched call", () => {